Handle getting config values exceptions

+++ AimpLyrics/Settings/ILyricsPluginSettings.cs
@@ -4,6 +4,6 @@
     {
         bool OpenWindowOnInitializing { get; set; }
         bool RestoreWindowHeight { get; set; }
-        int WindowHeight { get; set; }
+        double WindowHeight { get; set; }
     }
 }
